完善收取QKC逻辑 修改整点福利页面 修改integral数据类型 调试抢红包相关接口

+++ app/src/main/java/com/sbai/bcnews/model/mine/MyIntegral.java
@@ -12,7 +12,7 @@ public class MyIntegral {
      */
 
     private int rate;
-    private int integral;
+    private double integral;
 
     public int getRate() {
         return rate;
@@ -22,11 +22,11 @@ public class MyIntegral {
         this.rate = rate;
     }
 
-    public int getIntegral() {
+    public double getIntegral() {
         return integral;
     }
 
-    public void setIntegral(int integral) {
+    public void setIntegral(double integral) {
         this.integral = integral;
     }
 }
